requirements.txt
langchain-astradb
langchain 
langchain-openai
langchain-huggingface
datasets
pypdf
python-dotenv
flask

-e .